requirements.txt
python-telegram-bot>=21.0
requests>=2.30
APScheduler>=3.10